# Conflicts:
#	irs-api/src/test/java/net/catenax/irs/IrsFunctionalTest.java

+++ irs-models/src/main/java/net/catenax/irs/component/Quantity.java
@@ -21,7 +21,7 @@ import lombok.extern.jackson.Jacksonized;
 @Builder(toBuilder = true)
 public class Quantity {
 
-    private Integer quantityNumber;
+    private Double quantityNumber;
 
     private MeasurementUnit measurementUnit;
 }
